Broke main code into 2 files.
