+++ MainGUICode/AdditionalPages.py
@@ -1,0 +1,48 @@
+import tkinter as tkinter
+import tkinter.font as tkFont
+import MainGUI_WithKeypadEdit as Main
+
+class SpecificSearch(tkinter.Frame):
+    def __init__(self, master=None):
+        self.master = master
+        tkinter.Frame.__init__(self, master)
+        tkinter.Frame.configure(self,bg='')
+        tkinter.Label(self, bg='Green',text="Specified Search", font=('Helvetica', 18, "bold")).grid(row=0)
+
+        self.homeButton = tkinter.Button(self, text="Home",
+            command=lambda: master.switch_frame(Main.StartPage))
+        self.homeButton.grid(row=2, column=0, sticky="W"+"E")
+
+class RecentSearches(tkinter.Frame):
+    def __init__(self, master=None):
+        self.master = master
+        tkinter.Frame.__init__(self, master)
+        tkinter.Frame.configure(self,bg='')
+        tkinter.Label(self, bg='Green',text="Recent Searches", font=('Helvetica', 18, "bold")).grid(row=0)
+
+        self.homeButton = tkinter.Button(self, text="Home",
+            command=lambda: master.switch_frame(Main.StartPage))
+        self.homeButton.grid(row=2, column=0, sticky="W"+"E")
+
+class QuickSearch(tkinter.Frame):
+    def __init__(self, master=None):
+        self.master = master
+        tkinter.Frame.__init__(self, master)
+        tkinter.Frame.configure(self,bg='')
+        tkinter.Label(self, width = 35,bg='Green',text="Quick Search", font=('Helvetica', 18, "bold")).grid(row=0, columnspan=3)
+        L2fontStyle = tkFont.Font(family = "Lucida Grande", size =10)
+
+        self.homeButton = tkinter.Button(self, text="Home",
+            command=lambda: master.switch_frame(Main.StartPage))
+        self.homeButton.grid(row=10, column=1, sticky="W"+"E")
+
+        self.Resistors= tkinter.Button(self, text="Resistors", width=20,font = L2fontStyle,
+                           height=2, command=lambda: master.switch_frame(Main.StartPage))
+        self.Resistors.grid(row=1, column=0, pady=2)
+        self.Inductors= tkinter.Button(self, text="Inductors", width=20,font = L2fontStyle,
+                           height=2, command=lambda: master.switch_frame(Main.StartPage))
+        self.Inductors.grid(row=2, column=0, pady=2)
+
+        self.Capacitors= tkinter.Button(self, text="Capacitors", width=20,font = L2fontStyle,
+                           height=2, command=lambda: master.switch_frame(Main.StartPage))
+        self.Capacitors.grid(row=1, column=2, pady=2)
